Replace compression and decompression functions and update save/restore delay
Previous compression function took about 4 seconds to complete. This meant that we should have 5 seconds delay for save operation.
I replaced this slow compression and decompression algorithm with the faster one. It takes about 0.4 seconds to do the same compression that took 4 seconds previously.
This means that we can decrease the save/restore delay to 1500 seconds safely
But for faster loads and local development I put 700 and it works just fine, I put 1500 to be on a safe side, like for cases with 70 tabs or something.